s3-dcerpc: fix build warning seen with -O3.
"warning: assuming signed overflow does not occur when assuming that (X + c) < X is always false"

Guenther
(cherry picked from commit bb8936c08709ea5e59eeba2f67359c7511fb89f0)

+++ source3/librpc/rpc/dcerpc_helpers.c
@@ -898,9 +898,9 @@ NTSTATUS dcerpc_check_auth(struct pipe_auth_data *auth,
 	if (pkt->auth_length > pkt->frag_length) {
 		return NT_STATUS_INFO_LENGTH_MISMATCH;
 	}
-	if ((pkt->auth_length
-	     + DCERPC_AUTH_TRAILER_LENGTH < pkt->auth_length) ||
-	    (pkt->auth_length
+	if (((unsigned int)pkt->auth_length
+	     + DCERPC_AUTH_TRAILER_LENGTH < (unsigned int)pkt->auth_length) ||
+	    ((unsigned int)pkt->auth_length
 	     + DCERPC_AUTH_TRAILER_LENGTH < DCERPC_AUTH_TRAILER_LENGTH)) {
 		/* Integer wrap attempt. */
 		return NT_STATUS_INFO_LENGTH_MISMATCH;
